pyproject.toml
[tool.poetry]
name = "meraki"
version = "0.1.0"
description = "An ecommerce store"
authors = ["SphericalKat <amolele@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
uvicorn = "^0.17.4"
gunicorn = "^20.1.0"
Pillow = "^9.0.1"
django-storages = "^1.12.3"
boto3 = "^1.21.2"
django-environ = "^0.8.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
